requirements.txt
sentence-transformers
transformers
torch
faiss-cpu
streamlit
jsonlines
datasets
tqdm
flask
flask-cors